feat: Add Performer

+++ 5weeks/refator string calculator/gradle/src/main/java/string-calculator/domain/Calculator.java
@@ -11,7 +11,7 @@ public class Calculator {
         return a * b;
     }
 
-    public static double divNum(int a, int b) {
+    public static int divNum(int a, int b) {
         try {
             return a / b;
         }
